requirements.txt
biopython==1.81
tensorflow==2.13.1
scikit-learn==1.2.2
numpy==1.24.3
matplotlib==3.7.1
scipy==1.10.1